+++ R/fitDistBayes.R
@@ -1,63 +1,150 @@
 #' Fit the plain zeta model with the MCMC Bayesian implementation.
 #'
-#' @param x An object of class `psData`.
+#' @param x An input object or numeric vector required by the helper.
 #' @param prior A prior object created by `makePrior()`.
 #' @param nterms Number of fitted P/S probability terms to retain.
-#' @param ... MCMC controls passed to the posterior engine, including `shape0`,
-#'   `nIter`, `nBurnIn`, and `silent`.
+#' @param ... Additional arguments passed to the underlying fitting or helper routine.
 #' @return A Bayesian `psFit` object.
 #' @keywords internal
 #' @noRd
-fitDistBayes = function(x, prior = makePrior(), nterms, ...) {
-  if (!is(x, "psData")) {
+fitDistBayes = function(x, prior = makePrior(), nterms, ...){
+
+  nvals = 1:nterms
+  if(!is(x, "psData")){
     stop("x must be an object of class psData")
   }
 
-  model = zetaModel()
-  engine = mcmcPosteriorEngine()
-  representation = fitPosterior(
-    engine = engine,
-    model = model,
-    x = x,
-    prior = prior,
-    ...
-  )
-  summary = summarisePosterior(engine, model, representation)
-  pointEstimate = posteriorPointEstimate(engine, model, representation)
-
-  shape = unname(pointEstimate["shape"])
-  varShape = summary$sd[summary$parameter == "shape"]^2
-  probabilityIndices = if (x$type == "P") {
-    0:(nterms - 1L)
-  } else {
-    seq_len(nterms)
+  if(!is(x, "psData")){
+    stop("x must be an object of class psData")
   }
-  fittedMatrix = modelProbabilities(
-    model = model,
-    parameters = pointEstimate,
-    n = probabilityIndices,
-    type = x$type
-  )
-  fitted = as.numeric(fittedMatrix[1L, ])
-  names(fitted) = colnames(fittedMatrix)
 
-  chain = representation$value$chain
-  densityEstimate = density(chain, from = prior$range[1])
-  pdf = splinefun(densityEstimate$x, densityEstimate$y)
+  ## check which arguments have been provided and provide defaults
+  ## if they haven't
+
+  dotargs = list(...)
+  is.arg = function(arg){
+    return(tolower(arg) %in% tolower(names(dotargs)))
+  }
+
+  shape0 = ifelse(is.arg("shape0"), dotargs$shape0, 2)
+  validateZetaShape(shape0, "shape0")
+
+  nIter = ifelse(is.arg("nIter"), dotargs$nIter, 1e4)
+  nBurnIn = ifelse(is.arg("nBurnIn"), dotargs$nBurnIn, 1e3)
+  silent = ifelse(is.arg("silent"), dotargs$silent, TRUE)
+
+  if(nIter < 1000){
+    warning("The number of samples from the MCMC chain really should be 1000 or higher.")
+  }
+
+  if(nIter <= 0 || nBurnIn <=0){
+    stop("nIter and nBurnIn must be greater than zero.")
+  }
+
+  a = prior$range[1]
+  b = prior$range[2]
+
+  if(b <= a){
+    stop("b must be greater than a!")
+  }
+
+  W = b - a
+  if(W <= 0){
+    stop("This should never happen.")
+  }
+
+  if(length(x$data$n) < 2){
+    if(x$type == "S"){
+      stop("There has to be at least one value higher than 1")
+    }else{
+      stop("There has to be at least one value higher than 0")
+    }
+  }
+
+  obsData = if(x$type == 'P'){ ## the main difference is that the values need 1 added
+                x$data$n + 1
+            }else{
+              x$data$n
+            }
+
+  logLik = function(shape){
+    #-sum(VGAM::dzeta(rep(obsData, x$data$rn), shape = shape, log = TRUE))
+    sum(x$data$rn * dzetaStandard(obsData, shape = shape, log = TRUE))
+  }
+
+  nTotal = nIter + nBurnIn
+
+  draws = runif(nTotal, a, b)
+
+  chain = numeric(nIter)
+  log.u = log(runif(nTotal))
+
+  priorLogd = prior$logd
+  ll0 = logLik(shape0) + priorLogd(shape0)
+
+  if (!is.finite(ll0)){
+    stop("Log likelihood is not finite at starting value")
+  }
+
+  if(!silent){
+    pb = txtProgressBar(1, nTotal, 1, style = 3, label = 'Burning in')
+  }
+
+  i = 1
+  while(i <= nTotal){
+
+    shape1 = draws[i]
+    ll1 = logLik(shape1) + priorLogd(shape1)
+
+    if(ll1 > ll0 || log.u[i] < (ll1 - ll0)){
+      shape0 = shape1
+      ll0 = ll1
+    }
+
+    if(i > nBurnIn){
+      chain[i - nBurnIn] = shape0
+    }
+    i = i + 1
+    if(!silent){
+      if(i <= nBurnIn){
+        setTxtProgressBar(pb, i)
+      }else{
+        setTxtProgressBar(pb, i, label = 'Sampling')
+      }
+    }
+  }
+
+  if(!silent){
+    close(pb)
+  }
+
+  fit = list()
+  fit$par = shape = mean(chain)
+  var.shape = var(chain)
+
+  fitted = dzetaStandard(nvals, shape = shape)
+  names(fitted) = if(x$type == 'P'){
+    paste0("P", nvals - 1)
+  }else{
+    paste0("S", nvals)
+  }
+
+  d = density(chain, from = a)
+  pdf = splinefun(d$x, d$y)
 
   result = list(
     psData = x,
-    fit = list(par = shape),
+    fit = fit,
     shape = shape,
-    var.shape = varShape,
+    var.shape = var.shape,
     fitted = fitted,
     chain = chain,
     pdf = pdf,
-    posteriorRepresentation = representation,
     model = "zeta",
     method = "bayes"
   )
 
   class(result) = "psFit"
-  result
+
+  return(result)
 }